miq/tests/requirements.txt
coverage==5.5
django-environ==0.4.5
djangorestframework==3.12.4
pytest==6.2.4
pytest-cov==2.11.1
pytest-django==4.2.0
selenium==3.141.0